backend/requirements.txt
requests
beautifulsoup4
httpx
httpx[http2]
chromadb
llama-index
llama-index-vector-stores-chroma
llama-index-embeddings-google-genai
llama-index-llms-google-genai
google-generativeai
google-genai
python-dotenv
fastapi
uvicorn